refactor: update Mediator usage to FS.Mediator and improve method names for clarity
