+++ src/FeatureBasedFolderStructure.Application/Common/Behaviors/AuthorizationBehavior.cs
@@ -1,7 +1,7 @@
 using System.Reflection;
 using FeatureBasedFolderStructure.Application.Common.Attributes;
 using FS.AspNetCore.ResponseWrapper.Exceptions;
-using MediatR;
+using FS.Mediator.Features.RequestHandling.Core;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
 
@@ -12,17 +12,17 @@ public class AuthorizationBehavior<TRequest, TResponse>(
     IHttpContextAccessor httpContextAccessor)
     : IPipelineBehavior<TRequest, TResponse> where TRequest : notnull
 {
-    public async Task<TResponse> Handle(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
+    public async Task<TResponse> HandleAsync(TRequest request, RequestHandlerDelegate<TResponse> next, CancellationToken cancellationToken)
     {
         var requestType = request.GetType();
 
         // RequiresClaim attribute'unu kontrol et
-        var requiresClaimAttributes = requestType.GetCustomAttributes<RequiresClaimAttribute>(true);
+        var requiresClaimAttributes = requestType.GetCustomAttributes<RequiresClaimAttribute>(true).ToList();
 
         // Eğer attribute yoksa, yetkilendirme kontrolü gerekmez
-        if (!requiresClaimAttributes.Any())
+        if (requiresClaimAttributes.Count == 0)
         {
-            return await next();
+            return await next(cancellationToken);
         }
 
         logger.LogInformation("Authorization kontrolü yapılıyor: {RequestType}", requestType.Name);
@@ -42,26 +42,22 @@ public class AuthorizationBehavior<TRequest, TResponse>(
             if (attribute.RequireAllValues)
             {
                 // Tüm değerlerin mevcut olması gerekiyorsa
-                if (!attribute.ClaimValues.All(value =>
-                        user.HasClaim(c => c.Type == attribute.ClaimType && c.Value == value)))
-                {
-                    logger.LogWarning("Erişim reddedildi: Kullanıcı gerekli tüm claim değerlerine sahip değil. Claim: {ClaimType}", attribute.ClaimType);
-                    throw new ForbiddenAccessException($"Bu işlemi gerçekleştirmek için gereken '{attribute.ClaimType}' yetkisine sahip değilsiniz.");
-                }
+                if (attribute.ClaimValues.All(value =>
+                        user.HasClaim(c => c.Type == attribute.ClaimType && c.Value == value))) continue;
+                logger.LogWarning("Erişim reddedildi: Kullanıcı gerekli tüm claim değerlerine sahip değil. Claim: {ClaimType}", attribute.ClaimType);
             }
             else
             {
                 // En az bir değerin mevcut olması yeterli
-                if (!attribute.ClaimValues.Any(value =>
-                        user.HasClaim(c => c.Type == attribute.ClaimType && c.Value == value)))
-                {
-                    logger.LogWarning("Erişim reddedildi: Kullanıcı gerekli claim değerlerine sahip değil. Claim: {ClaimType}", attribute.ClaimType);
-                    throw new ForbiddenAccessException($"Bu işlemi gerçekleştirmek için gereken '{attribute.ClaimType}' yetkisine sahip değilsiniz.");
-                }
+                if (attribute.ClaimValues.Any(value =>
+                        user.HasClaim(c => c.Type == attribute.ClaimType && c.Value == value))) continue;
+                logger.LogWarning("Erişim reddedildi: Kullanıcı gerekli claim değerlerine sahip değil. Claim: {ClaimType}", attribute.ClaimType);
             }
+
+            throw new ForbiddenAccessException($"Bu işlemi gerçekleştirmek için gereken '{attribute.ClaimType}' yetkisine sahip değilsiniz.");
         }
 
         logger.LogInformation("Yetkilendirme başarılı: {RequestType}", requestType.Name);
-        return await next();
+        return await next(cancellationToken);
     }
 }
