+++ src/FeatureBasedFolderStructure.Application/FeatureBasedFolderStructure.Application.csproj
@@ -12,7 +12,6 @@
       <PackageReference Include="FluentValidation.AspNetCore" Version="11.3.1" />
       <PackageReference Include="FS.AspNetCore.ResponseWrapper" Version="9.0.0" />
       <PackageReference Include="FS.AutoServiceDiscovery.Extensions" Version="9.0.0" />
-      <PackageReference Include="MediatR" Version="12.5.0" />
       <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.6" />
     </ItemGroup>
