+++ src/FeatureBasedFolderStructure.Application/Features/v1/Auth/Commands/ChangePassword/ChangePasswordCommand.cs
@@ -1,9 +1,21 @@
-using MediatR;
+using FeatureBasedFolderStructure.Application.Common.Interfaces;
+using FeatureBasedFolderStructure.Application.Interfaces.Users;
+using FS.Mediator.Features.RequestHandling.Core;
 
 namespace FeatureBasedFolderStructure.Application.Features.v1.Auth.Commands.ChangePassword;
 
-public class ChangePasswordCommand : IRequest<Unit>
+public sealed record ChangePasswordCommand(string CurrentPassword, string NewPassword) : IRequest<Unit>;
+
+internal class ChangePasswordCommandHandler(IApplicationUserService applicationUserService, ICurrentUserService currentUserService)
+    : IRequestHandler<ChangePasswordCommand, Unit>
 {
-    public string CurrentPassword { get; set; } = null!;
-    public string NewPassword { get; set; } = null!;
+    public async Task<Unit> HandleAsync(ChangePasswordCommand request, CancellationToken cancellationToken)
+    {
+        if (!Guid.TryParse(currentUserService.UserId, out var userId))
+            throw new UnauthorizedAccessException("Kullanıcı bilgisi alınamadı.");
+
+        var user = await applicationUserService.GetByIdAsync(userId, cancellationToken);
+        await applicationUserService.ChangePasswordAsync(user.Id, request.CurrentPassword, request.NewPassword, cancellationToken);
+        return Unit.Value;
+    }
 }
