+++ src/FeatureBasedFolderStructure.Application/Features/v1/Categories/Commands/UpdateCategory/UpdateCategoryCommand.cs
@@ -1,10 +1,33 @@
-using MediatR;
+using FeatureBasedFolderStructure.Domain.Entities.Catalogs;
+using FS.AspNetCore.ResponseWrapper.Exceptions;
+using FS.EntityFramework.Library.UnitOfWorks;
+using FS.Mediator.Features.RequestHandling.Core;
+using Microsoft.Extensions.Logging;
 
 namespace FeatureBasedFolderStructure.Application.Features.v1.Categories.Commands.UpdateCategory;
 
-public class UpdateCategoryCommand : IRequest<Unit>
+public sealed record UpdateCategoryCommand(int Id, string Name, string? Description) : IRequest<Unit>;
+
+internal class UpdateCategoryCommandHandler(
+    IUnitOfWork unitOfWork,
+    ILogger<UpdateCategoryCommandHandler> logger)
+    : IRequestHandler<UpdateCategoryCommand, Unit>
 {
-    public int Id { get; set; }
-    public string Name { get; set; } = null!;
-    public string? Description { get; set; }
+    public async Task<Unit> HandleAsync(UpdateCategoryCommand request, CancellationToken cancellationToken)
+    {
+        var categoryRepository = unitOfWork.GetRepository<Category, int>();
+        var entity = await categoryRepository.GetByIdAsync(request.Id, cancellationToken: cancellationToken);
+
+        if (entity == null)
+            throw new NotFoundException(nameof(Category), request.Id);
+
+        entity.Name = request.Name;
+        entity.Description = request.Description;
+
+        await categoryRepository.UpdateAsync(entity, cancellationToken: cancellationToken);
+        await unitOfWork.SaveChangesAsync(cancellationToken); 
+        logger.LogInformation("Updated Category {CategoryId}", entity.Id);
+
+        return Unit.Value;
+    }
 }
