+++ src/FeatureBasedFolderStructure.Application/Features/v1/Categories/Queries/GetCategoryDetail/GetCategoryDetailQuery.cs
@@ -1,9 +1,24 @@
+using AutoMapper;
 using FeatureBasedFolderStructure.Application.Features.v1.Categories.DTOs;
-using MediatR;
+using FeatureBasedFolderStructure.Domain.Entities.Catalogs;
+using FS.AspNetCore.ResponseWrapper.Exceptions;
+using FS.EntityFramework.Library.UnitOfWorks;
+using FS.Mediator.Features.RequestHandling.Core;
 
 namespace FeatureBasedFolderStructure.Application.Features.v1.Categories.Queries.GetCategoryDetail;
 
-public class GetCategoryDetailQuery:IRequest<CategoryDto>
+public sealed record GetCategoryDetailQuery(int Id) : IRequest<CategoryDto>;
+
+internal class GetCategoryDetailQueryHandler(
+    IUnitOfWork unitOfWork,
+    IMapper mapper) : IRequestHandler<GetCategoryDetailQuery, CategoryDto>
 {
-    public int Id { get; set; }
+    public async Task<CategoryDto> HandleAsync(GetCategoryDetailQuery request, CancellationToken cancellationToken)
+    {
+        var categoryRepository = unitOfWork.GetRepository<Category, int>();
+        var category = await categoryRepository.GetByIdAsync(request.Id, true, cancellationToken);
+        if (category == null)
+            throw new NotFoundException(nameof(Category), request.Id);
+        return mapper.Map<CategoryDto>(category);
+    }
 }
