+++ src/FeatureBasedFolderStructure.Application/Features/v1/Products/Commands/CreateProduct/CreateProductCommand.cs
@@ -1,12 +1,41 @@
+using AutoMapper;
 using FeatureBasedFolderStructure.Application.Features.v1.Products.DTOs;
-using MediatR;
+using FeatureBasedFolderStructure.Application.Features.v1.Products.Rules;
+using FeatureBasedFolderStructure.Domain.Entities.Catalogs;
+using FS.EntityFramework.Library.UnitOfWorks;
+using FS.Mediator.Features.RequestHandling.Core;
+using Microsoft.Extensions.Logging;
 
 namespace FeatureBasedFolderStructure.Application.Features.v1.Products.Commands.CreateProduct;
 
-public class CreateProductCommand : IRequest<ProductDto>
+public sealed record CreateProductCommand(string Name, decimal Price, string? Description, int CategoryId) : IRequest<ProductDto>;
+
+internal class CreateProductCommandHandler(
+    IUnitOfWork unitOfWork,
+    ProductBusinessRules productBusinessRules,
+    ILogger<CreateProductCommandHandler> logger,
+    IMapper mapper)
+    : IRequestHandler<CreateProductCommand, ProductDto>
 {
-    public string Name { get; set; } = null!;
-    public decimal Price { get; set; }
-    public string? Description { get; set; }
-    public int CategoryId { get; set; }
+    public async Task<ProductDto> HandleAsync(CreateProductCommand request, 
+        CancellationToken cancellationToken)
+    {
+        await productBusinessRules.CheckIfCategoryExists(request.CategoryId, cancellationToken);
+        
+        var entity = new Product
+        {
+            Name = request.Name,
+            Price = request.Price,
+            Description = request.Description,
+            CategoryId = request.CategoryId
+        };
+        entity.UpdatePrice(request.Price, "TRY");
+        
+        var productRepository = unitOfWork.GetRepository<Product, int>();
+        await productRepository.AddAsync(entity, cancellationToken: cancellationToken);
+        await unitOfWork.SaveChangesAsync(cancellationToken);
+        logger.LogInformation("Created Product {ProductId}", entity.Id);
+
+        return mapper.Map<ProductDto>(entity);
+    }
 }
