+++ src/FeatureBasedFolderStructure.Application/Features/v1/Products/Commands/UpdateProduct/UpdateProductCommand.cs
@@ -1,12 +1,42 @@
-using MediatR;
+using FeatureBasedFolderStructure.Application.Features.v1.Products.Rules;
+using FeatureBasedFolderStructure.Domain.Entities.Catalogs;
+using FS.AspNetCore.ResponseWrapper.Exceptions;
+using FS.EntityFramework.Library.UnitOfWorks;
+using FS.Mediator.Features.RequestHandling.Core;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Logging;
 
 namespace FeatureBasedFolderStructure.Application.Features.v1.Products.Commands.UpdateProduct;
 
-public class UpdateProductCommand : IRequest<Unit>
+public sealed record UpdateProductCommand(int Id, string Name, decimal Price, string? Description, int CategoryId) : IRequest<Unit>;
+
+internal class UpdateProductCommandHandler(
+    IUnitOfWork unitOfWork,
+    ProductBusinessRules productBusinessRules,
+    ILogger<UpdateProductCommandHandler> logger)
+    : IRequestHandler<UpdateProductCommand, Unit>
 {
-    public int Id { get; set; }
-    public string Name { get; set; } = null!;
-    public decimal Price { get; set; }
-    public string? Description { get; set; }
-    public int CategoryId { get; set; }
+    public async Task<Unit> HandleAsync(UpdateProductCommand request, CancellationToken cancellationToken)
+    {
+        var productRepository = unitOfWork.GetRepository<Product, int>();
+        var entity = await productRepository.GetQueryable().Include(p => p.CurrentPrice).FirstOrDefaultAsync(p => p.Id == request.Id, cancellationToken: cancellationToken);
+
+        if (entity == null)
+            throw new NotFoundException(nameof(Product), request.Id);
+
+        await productBusinessRules.CheckIfCategoryExists(request.CategoryId, cancellationToken);
+
+        entity.Name = request.Name;
+        entity.Price = request.Price;
+        entity.UpdatePrice(request.Price, "TRY");
+        entity.Description = request.Description;
+        entity.CategoryId = request.CategoryId;
+
+        await productRepository.UpdateAsync(entity, cancellationToken: cancellationToken);
+        await unitOfWork.SaveChangesAsync(cancellationToken);
+        
+        logger.LogInformation("Updated Product {ProductId}", entity.Id);
+
+        return Unit.Value;
+    }
 }
