+++ src/FeatureBasedFolderStructure.Application/Features/v1/Products/Queries/GetProducts/GetProductsQuery.cs
@@ -1,11 +1,27 @@
+using AutoMapper;
 using FeatureBasedFolderStructure.Application.Features.v1.Products.DTOs;
 using FeatureBasedFolderStructure.Application.Requests;
 using FeatureBasedFolderStructure.Domain.Common.Paging;
-using MediatR;
+using FeatureBasedFolderStructure.Domain.Entities.Catalogs;
+using FS.EntityFramework.Library.UnitOfWorks;
+using FS.Mediator.Features.RequestHandling.Core;
 
 namespace FeatureBasedFolderStructure.Application.Features.v1.Products.Queries.GetProducts;
 
-public class GetProductsQuery : IRequest<PagedResult<ProductDto>>
+public sealed record GetProductsQuery(PageRequest PageRequest) : IRequest<PagedResult<ProductDto>>;
+
+internal class GetProductsQueryHandler(
+    IUnitOfWork unitOfWork,
+    IMapper mapper) : IRequestHandler<GetProductsQuery, PagedResult<ProductDto>>
 {
-    public PageRequest PageRequest { get; set; } = new();
+    public async Task<PagedResult<ProductDto>> HandleAsync(GetProductsQuery request, CancellationToken cancellationToken)
+    {
+        var productRepository = unitOfWork.GetRepository<Product, int>();
+        var products = await productRepository.GetPagedAsync(
+            pageIndex: request.PageRequest.Page,
+            pageSize: request.PageRequest.PageSize,
+            orderBy: p => p.OrderBy(product => product.Name),
+            cancellationToken: cancellationToken);
+        return mapper.Map<PagedResult<ProductDto>>(products);
+    }
 }
