+++ src/FeatureBasedFolderStructure.Domain/FeatureBasedFolderStructure.Domain.csproj
@@ -8,7 +8,7 @@
 
     <ItemGroup>
       <PackageReference Include="FS.EntityFramework.Library" Version="9.0.6.6" />
-      <PackageReference Include="MediatR.Contracts" Version="2.0.1" />
+      <PackageReference Include="FS.Mediator" Version="9.0.0" />
       <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.6" />
       <PackageReference Include="Microsoft.EntityFrameworkCore.Abstractions" Version="9.0.6" />
     </ItemGroup>
