+++ src/FeatureBasedFolderStructure.Infrastructure/FeatureBasedFolderStructure.Infrastructure.csproj
@@ -7,7 +7,6 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="MediatR" Version="12.5.0" />
       <PackageReference Include="Microsoft.EntityFrameworkCore" Version="9.0.6" />
       <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="9.0.6">
         <PrivateAssets>all</PrivateAssets>
